task-06/requirements.txt
discord
beautifulsoup4
requests
python-dotenv
